Deprecate PanelController
`PanelController` was mainly used for changing the state of the panel,
i.e., to open/close/hide the panel, among other things. It used the
private animation controller to perform those operations.

The animation controller if made public or accessible from the
`PanelController` can be useful in the case of providing custom duration
and optional explicit control over the animation.

However, a change like that would make the function of PanelController
redundant.

Consequently, the limitation of the `SlidingUpPanel` to be "already
rendered" to use functions to control the state by using a `Widget`
reference separate from the main widget tree _is a bit of an underkill_
(is that even a word?).

This implementation _attempts_ to resolve this problem by the following
ways:
 - By making the `AnimationController` accessible and an optional
   constructor flag.
 - By making `SlidingUpPanelState` public.
 - Introducing **`SlidingUpPanel.of()`** to access the public state of
   the `SlidingUpPanel` similar to what is done in case of `Scaffold`
   and `ScaffoldState`
 - Make all "state controlling" methods (`open()`,`close(),`hide()`,
  `animateTo()`, etc) in the `SlidingUpPanelState` public.

Additionally,
 - `animationDuration`, `expandCurve`, `collapseCurve`, are added as flags
   to `SlidingUpPanel` constructor, which can override optional parameters
   in functions `animatePanelToPosition()`, `animatePanelToSnapPoint()` and
   `_onGestureEnd()`
 - The ScrollController has been properly disposed.
 - Documentation is added from the README.

Users will still be able to use PanelController without/with AnimationController.